Break up exercise item into exercise detail
